+++ R/PlotMap.R
@@ -168,7 +168,7 @@
 #'         useRaster = TRUE)
 #'
 #' r <- raster::raster(system.file("external/test.grd", package = "raster"))
-#' Pal <- colorspace::rainbow_hcl
+#' Pal <- function(n) viridisLite::magma(n, begin = 0.3)
 #' breaks <- seq(0, 2000, by = 200)
 #' PlotMap(r, breaks = breaks, pal = Pal, dms.tick = TRUE, bg.lines = TRUE,
 #'         scale.loc = "bottomright", contour.lines = list(col = "#1F1F1F"),
@@ -371,11 +371,15 @@ PlotMap <- function(r, layer=1, att=NULL, n=NULL, breaks=NULL,
     if (length(col) != n) stop("number of specified colors is incorrect")
     cols <- col
   } else {
-    if (is.function(pal)) {
+    if (n == 0) {
+      cols <- as.character()
+    } else if (is.function(pal)) {
       cols <- pal(n)
+    } else if (raster::is.factor(r)) {
+      cols <- GetTolColors(n)
     } else {
-      if (requireNamespace("colorspace", quietly=TRUE))
-        cols <- colorspace::rainbow_hcl(n, start=0.0, end=(360 * (n - 1) / n) * 0.8)
+      if (requireNamespace("viridisLite", quietly=TRUE))
+        cols <- viridisLite::viridis(n)
       else
         cols <- grDevices::rainbow(n, start=0.0, end=0.8)
     }
@@ -384,10 +388,9 @@ PlotMap <- function(r, layer=1, att=NULL, n=NULL, breaks=NULL,
 
   # plot color key
   if (draw.key & n > 0) {
-    is.categorical <- raster::is.factor(r)
     at <- if (is.null(labels$at)) at1 else labels$at
     if (is.null(labels$labels))
-      labels <- if (is.categorical) raster::levels(r)[[1]][, "att"] else TRUE
+      labels <- if (raster::is.factor(r)) raster::levels(r)[[1]][, "att"] else TRUE
     else
       labels <- labels$labels
     AddColorKey(mai=mai1, is.categorical=raster::is.factor(r),
@@ -577,7 +580,7 @@ PlotMap <- function(r, layer=1, att=NULL, n=NULL, breaks=NULL,
   }
 
   if (!is.null(scale.loc)) {
-    txt <- strsplit(proj4string(r), " ")[[1]]
+    txt <- strsplit(sp::proj4string(r), " ")[[1]]
     unit <- sub("^\\+units=", "", grep("^\\+units=", txt, value=TRUE))
     longlat <- "+proj=longlat" %in% txt
     AddScaleBar(unit=unit, longlat=longlat, loc=scale.loc)